+++ metrics/objects/transform/base.py
@@ -3,7 +3,8 @@ import numpy as np
 
 
 class Base:
+
     def parse_lnmk(self, item):
         for key in list(item.keys()):
-            item[key] = np.asarray(item[key])
+            item[key] = np.asarray(item[key]).astype(np.float32)
         return item
